requirements.txt
DrissionPage
ffmpeg-downloader
requests
openai-whisper
fake-useragent
psutil